+++ packages/document/template/common/button-size.vue
@@ -1,7 +1,7 @@
 <template>
 	<px-space>
 		<px-button size="small">Small</px-button>
-		<px-button>Default</px-button>
+		<px-button theme="success">Default</px-button>
 		<px-button size="large">Large</px-button>
 	</px-space>
 </template>