+++ packages/document/template/feedback/alert-type.vue
@@ -1,6 +1,7 @@
 <template>
 	<px-space direction="vertical">
 		<px-alert>Info</px-alert>
+		<px-alert type="notice">Notice</px-alert>
 		<px-alert type="success">Success</px-alert>
 		<px-alert type="warning">Warning</px-alert>
 		<px-alert type="error">Error</px-alert>